release 3.1.107 source code

+++ services/secmaster/src/main/java/com/huaweicloud/sdk/secmaster/v2/model/ListDataclassFieldsRequest.java
@@ -3,7 +3,6 @@ package com.huaweicloud.sdk.secmaster.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 /**
@@ -19,12 +18,12 @@ public class ListDataclassFieldsRequest {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "offset")
 
-    private BigDecimal offset;
+    private Integer offset;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "limit")
 
-    private BigDecimal limit;
+    private Integer limit;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "name")
@@ -68,7 +67,7 @@ public class ListDataclassFieldsRequest {
         this.workspaceId = workspaceId;
     }
 
-    public ListDataclassFieldsRequest withOffset(BigDecimal offset) {
+    public ListDataclassFieldsRequest withOffset(Integer offset) {
         this.offset = offset;
         return this;
     }
@@ -79,15 +78,15 @@ public class ListDataclassFieldsRequest {
      * maximum: 999999999
      * @return offset
      */
-    public BigDecimal getOffset() {
+    public Integer getOffset() {
         return offset;
     }
 
-    public void setOffset(BigDecimal offset) {
+    public void setOffset(Integer offset) {
         this.offset = offset;
     }
 
-    public ListDataclassFieldsRequest withLimit(BigDecimal limit) {
+    public ListDataclassFieldsRequest withLimit(Integer limit) {
         this.limit = limit;
         return this;
     }
@@ -98,11 +97,11 @@ public class ListDataclassFieldsRequest {
      * maximum: 100
      * @return limit
      */
-    public BigDecimal getLimit() {
+    public Integer getLimit() {
         return limit;
     }
 
-    public void setLimit(BigDecimal limit) {
+    public void setLimit(Integer limit) {
         this.limit = limit;
     }
